handle completed reviews without a made ECO
When we first started scraping LGBCE, the made ECO would always
be posted before the review is moved from current to complete.
It now seems to be relatively common practice for a review
to be moved to complete while only the draft ECO is posted.

So far I've been dealing with this by adding reviews
which violate this constraint to a list of exceptions.
Given this seems to be happening routinely,
at this point it makes sense to:

- Remove the validation rule requiring a complete review to have a made ECO
- Raise issue on the EE repo when the review moves from current to complete
- Pick up the legislation URL later when/if it gets posted